Split base_test.py into two more files
